GO-1080: Add exercise command
This change introduces cmd/exercise which can be used to fully
exercise the test bench. The script generates a matrix of all
unsafe routes and populates them with fake user data. There are
a number of small inconsistencies between route definitions which
prevents generalized logic, so there are separate generators for
each page.

The exercise command can be run from the root of the repo with:
  go run ./cmd/exercise -addr localhost:8080

The command uses the new Application-Framework header from the
response of the application root to determine what framework is
running.

An integration test is also added to run the script in pipelines.
This test asserts that known routes are exercised correctly but
does not recognize new untested routes automatically.
